# openapi.json
{
  "components": {
    "schemas": {
...
      "CommandEnvChannelResponse": {
        "properties": {
          "channel_id": {
            "title": "Channel Id",
            "type": "integer"
          },
          "guild_id": {
            "title": "Guild Id",
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/CommandEnvKind"
          },
          "owner_discord_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Discord Id"
          }
        },
        "required": [
          "guild_id",
          "channel_id",
          "kind"
        ],
        "title": "CommandEnvChannelResponse",
        "type": "object"
      },
      "CommandEnvKind": {
        "enum": [
          "admin_cmd",
          "tutor_cmd"
        ],
        "title": "CommandEnvKind",
        "type": "string"
      },
      "CommandEnvUpsertRequest": {
        "properties": {
          "channel_id": {
            "title": "Channel Id",
            "type": "integer"
          },
          "guild_id": {
            "title": "Guild Id",
            "type": "integer"
          },
          "kind": {
            "$ref": "#/components/schemas/CommandEnvKind"
          },
          "owner_discord_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Discord Id"
          }
        },
        "required": [
          "guild_id",
          "channel_id",
          "kind"
        ],
        "title": "CommandEnvUpsertRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Body of every non-2xx response (ADR 0006).",
        "properties": {
          "code": {
            "description": "Stable, machine-readable identifier in `snake_case`. Branch on this, never on `detail`.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "description": "Human-readable message. Safe to show; never contains internals.",
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Present only for request-validation failures: one entry per offending field.",
            "title": "Errors"
          }
        },
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FieldError": {
        "description": "One offending field of a request that failed validation.",
        "properties": {
          "loc": {
            "description": "Path to the field, starting with its location (`path`, `query`, `body`, ...).",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "message": {
            "description": "Human-readable reason the value was rejected.",
            "title": "Message",
            "type": "string"
          },
          "type": {
            "description": "Machine-readable identifier of the violated rule (e.g. `less_than_equal`).",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "message",
          "type"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2ClientCredentialsBearer": {
        "flows": {
          "clientCredentials": {
            "scopes": {
              "auth:clients:manage": "Manage application clients.",
              "bot:read": "Read bot API surface.",
              "bot:write": "Write bot API surface."
            },
            "tokenUrl": "/api/v1/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "Backend of the skill-platform",
    "title": "skillforge",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/bot/command-envs": {
      "put": {
        "operationId": "bot_upsert_command_env",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandEnvUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandEnvChannelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "invalid_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Invalid authentication credentials"
                    }
                  },
                  "missing_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Not authenticated"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "code": "forbidden",
                  "detail": "Not enough permissions"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing required scope: bot:write"
          },
          "409": {
            "content": {
              "application/json": {
                "examples": {
                  "command_env_conflict": {
                    "value": {
                      "code": "command_env_conflict",
                      "detail": "Owner already owns a command env of this kind in the guild"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Owner already owns a command env of this kind in the guild"
          },
          "422": {
            "content": {
              "application/json": {
                "examples": {
                  "command_env_validation": {
                    "value": {
                      "code": "command_env_validation",
                      "detail": "Command env references an unknown guild, channel, or owner"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Command env references an unknown guild, channel, or owner"
          }
        },
        "security": [
          {
            "OAuth2ClientCredentialsBearer": [
              "bot:write"
            ]
          }
        ],
        "summary": "Upsert Command Env Endpoint",
        "tags": [
          "bot"
        ]
      }
    },
    "/api/v1/bot/command-envs/{guild_id}/{channel_id}/{kind}": {
      "delete": {
        "operationId": "bot_delete_command_env",
        "parameters": [
          {
            "in": "path",
            "name": "guild_id",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Guild Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "channel_id",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Channel Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CommandEnvKind"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "invalid_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Invalid authentication credentials"
                    }
                  },
                  "missing_token": {
                    "value": {
                      "code": "unauthorized",
                      "detail": "Not authenticated"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "content": {
              "application/json": {
                "example": {
                  "code": "forbidden",
                  "detail": "Not enough permissions"
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing required scope: bot:write"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "command_env_not_found": {
                    "value": {
                      "code": "command_env_not_found",
                      "detail": "Command env channel not found"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Command env channel not found"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "code": "validation_error",
                  "detail": "Request validation failed",
                  "errors": [
                    {
                      "loc": [
                        "query",
                        "limit"
                      ],
                      "message": "Input should be less than or equal to 100",
                      "type": "less_than_equal"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "OAuth2ClientCredentialsBearer": [
              "bot:write"
            ]
          }
        ],
        "summary": "Delete Command Env Endpoint",
        "tags": [
          "bot"
        ]
      }
    },
...
  },
...
}